Refactor progress screen data and utilities
