+++ components/screens/progress/constants.ts
@@ -1,0 +1,11 @@
+export const PROGRESS_THEME = {
+  accent: "#E27D60",
+  rangeAccent: "#68A691",
+  surface: "#FFFFFF",
+  surfaceMuted: "#F8F6F3",
+  clear: "transparent",
+  border: "rgba(30,36,50,0.08)",
+  domainMenuShadow: "0 12px 24px -16px rgba(30,36,50,0.4)",
+  rangeShadowActive: "0 12px 22px -16px rgba(30,36,50,0.35)",
+  rangeShadowInactive: "0 2px 6px -4px rgba(30,36,50,0.18)",
+} as const;
